requirements.txt
asgiref==3.5.2
Django==4.1.4
django-cors-headers==3.13.0
django-token==0.1.3
djangorestframework==3.14.0
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
